ml_training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
mlflow>=2.7.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tqdm>=4.66.0
tensorboard>=2.14.0
